requirements.txt
google.cloud.language